removed all audioin stuff, still to separate them and deal with global variables
