+++ Urho3D/splash/processBuffer.h
@@ -1,0 +1,1 @@
+int processBuffer();
