+++ Urho3D/splash_build/CMakeFiles/Splash.dir/CXX.includecache
@@ -1276,6 +1276,34 @@ Urho3D/Resource/XMLFile.h
 Urho3D/IO/Log.h
 -
 
+/home/pi/sound-pirates/Urho3D/splash/fft.cpp
+RtAudio.h
+/home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+fftw3.h
+-
+iostream
+-
+cstdlib
+-
+math.h
+-
+SFML/Graphics.hpp
+-
+cstring
+-
+stdio.h
+-
+SFML/Audio.hpp
+-
+vector
+-
+complex
+-
+string
+-
+lol.h
+/home/pi/sound-pirates/Urho3D/splash/lol.h
+
 /home/pi/sound-pirates/Urho3D/splash/fft.h
 
 /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
@@ -1283,6 +1311,70 @@ Urho3D/IO/Log.h
 /home/pi/sound-pirates/Urho3D/splash/lol.h
 
 /home/pi/sound-pirates/Urho3D/splash/playNote.h
+
+/home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
+RtAudio.h
+/home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+fftw3.h
+-
+iostream
+-
+cstdlib
+-
+math.h
+-
+SFML/Graphics.hpp
+-
+cstring
+-
+stdio.h
+-
+SFML/Audio.hpp
+-
+vector
+-
+complex
+-
+string
+-
+lol.h
+/home/pi/sound-pirates/Urho3D/splash/lol.h
+fft.h
+/home/pi/sound-pirates/Urho3D/splash/fft.h
+playNote.h
+/home/pi/sound-pirates/Urho3D/splash/playNote.h
+instructionsStatements.h
+/home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
+sys/types.h
+-
+stdlib.h
+-
+fcntl.h
+-
+string
+-
+memory
+-
+fstream
+-
+sstream
+-
+iostream
+-
+stdio.h
+-
+unistd.h
+-
+poll.h
+-
+thread
+-
+chrono
+-
+signal.h
+-
+
+/home/pi/sound-pirates/Urho3D/splash/processBuffer.h
 
 /home/pi/sound-pirates/Urho3D/splash/splash.cpp
 RtAudio.h
@@ -1315,6 +1407,10 @@ fft.h
 /home/pi/sound-pirates/Urho3D/splash/fft.h
 playNote.h
 /home/pi/sound-pirates/Urho3D/splash/playNote.h
+instructionsStatements.h
+/home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
+processBuffer.h
+/home/pi/sound-pirates/Urho3D/splash/processBuffer.h
 sys/types.h
 -
 stdlib.h
@@ -1421,8 +1517,6 @@ splash.h
 /home/pi/sound-pirates/Urho3D/splash/splash.h
 Urho3D/DebugNew.h
 -
-instructionsStatements.h
-/home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
 
 /home/pi/sound-pirates/Urho3D/splash/splash.h
 Sample.h
