+++ Urho3D/splash_build/CMakeFiles/Splash.dir/cmake_clean.cmake
@@ -3,6 +3,7 @@ file(REMOVE_RECURSE
   "CMakeFiles/Splash.dir/fft.cpp.o"
   "CMakeFiles/Splash.dir/instructionsStatements.cpp.o"
   "CMakeFiles/Splash.dir/playNote.cpp.o"
+  "CMakeFiles/Splash.dir/processBuffer.cpp.o"
   "CMakeFiles/Splash.dir/realtime-beat.cpp.o"
   "CMakeFiles/Splash.dir/splash.cpp.o"
   "bin/Splash"
