+++ Urho3D/splash_build/CMakeFiles/Splash.dir/depend.internal
@@ -20,6 +20,13 @@ CMakeFiles/Splash.dir/playNote.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.cpp
  /home/pi/sound-pirates/Urho3D/splash/sinewave.h
+CMakeFiles/Splash.dir/processBuffer.cpp.o
+ /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+ /home/pi/sound-pirates/Urho3D/splash/fft.h
+ /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
+ /home/pi/sound-pirates/Urho3D/splash/lol.h
+ /home/pi/sound-pirates/Urho3D/splash/playNote.h
+ /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
 CMakeFiles/Splash.dir/realtime-beat.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
  /home/pi/sound-pirates/Urho3D/splash/lol.h
@@ -179,5 +186,6 @@ CMakeFiles/Splash.dir/splash.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
  /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.h
+ /home/pi/sound-pirates/Urho3D/splash/processBuffer.h
  /home/pi/sound-pirates/Urho3D/splash/splash.cpp
  /home/pi/sound-pirates/Urho3D/splash/splash.h
