+++ Urho3D/splash_build/CMakeFiles/Splash.dir/depend.make
@@ -20,6 +20,13 @@ CMakeFiles/Splash.dir/playNote.cpp.o: /home/pi/sound-pirates/Urho3D/splash/lol.h
 CMakeFiles/Splash.dir/playNote.cpp.o: /home/pi/sound-pirates/Urho3D/splash/playNote.cpp
 CMakeFiles/Splash.dir/playNote.cpp.o: /home/pi/sound-pirates/Urho3D/splash/sinewave.h
 
+CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/fft.h
+CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
+CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/lol.h
+CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/playNote.h
+CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
+
 CMakeFiles/Splash.dir/realtime-beat.cpp.o: /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
 CMakeFiles/Splash.dir/realtime-beat.cpp.o: /home/pi/sound-pirates/Urho3D/splash/lol.h
 CMakeFiles/Splash.dir/realtime-beat.cpp.o: /home/pi/sound-pirates/Urho3D/splash/realtime-beat.cpp
@@ -178,6 +185,7 @@ CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/fft.h
 CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
 CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/lol.h
 CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/playNote.h
+CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/processBuffer.h
 CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/splash.cpp
 CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/splash.h
