+++ Urho3D/splash_build/CMakeFiles/Splash.dir/progress.make
@@ -5,4 +5,5 @@ CMAKE_PROGRESS_4 = 4
 CMAKE_PROGRESS_5 = 5
 CMAKE_PROGRESS_6 = 6
 CMAKE_PROGRESS_7 = 7
+CMAKE_PROGRESS_8 = 8
